requirements.txt
requests
argparse
python-dotenv
clean-text
PyMuPDF
pinecone-client